Address warnings (cont.)

+++ dds/DCPS/Serializer.inl
@@ -408,7 +408,7 @@ Serializer::skip(size_t n, int size)
     return false;
   }
 
-  for (size_t len = static_cast<size_t>(n * size); len;) {
+  for (size_t len = n * static_cast<size_t>(size); len;) {
     if (!current_) {
       good_bit_ = false;
       return false;
@@ -425,7 +425,7 @@ Serializer::skip(size_t n, int size)
   }
 
   if (good_bit_) {
-    rpos_ += n * size;
+    rpos_ += n * static_cast<size_t>(size);
   }
   return good_bit();
 }
@@ -783,7 +783,7 @@ bool Serializer::align_r(size_t al)
   }
   al = (std::min)(al, encoding().max_align());
   const size_t len =
-    (al - ptrdiff_t(current_->rd_ptr()) + align_rshift_) % al;
+    (al - static_cast<size_t>(current_->rd_ptr()) + align_rshift_) % al;
 
   return skip(static_cast<ACE_CDR::UShort>(len));
 }
@@ -800,7 +800,7 @@ bool Serializer::align_w(size_t al)
   }
   al = (std::min)(al, encoding().max_align());
   size_t len =
-    (al - ptrdiff_t(current_->wr_ptr()) + align_wshift_) % al;
+    (al - static_cast<size_t>(current_->wr_ptr()) + align_wshift_) % al;
   while (len) {
     if (!current_) {
       good_bit_ = false;
